directory: Document class and functions
