+++ src/directory/index.ts
@@ -1,36 +1,108 @@
+/**
+ * This module defines the [[Directory]] class.
+ * @module
+ */
+
 import type { Block } from "../block";
-import { decomposeBlock } from "../block";
 import type { BlockId } from "../block/blockId";
 
-class Directory {
-  blocks: Record<string, Block>;
-
-  constructor() {
-    this.blocks = {};
+const decomposeBlock = (
+  block: Block,
+  prefix?: string
+): [Block, Record<string, Block>] => {
+  if (prefix === undefined || prefix === null) {
+    prefix = "";
   }
 
+  const { name, ay, assign, match, satisfy, url, info, ...subblocks } = block;
+  const mainBlock = { name, ay, assign, match, satisfy, url, info };
+  const flatSubblocks = Object.entries(subblocks)
+    .map(([sbName, sb]) => {
+      const sbPrefix = [prefix, sbName].join("/");
+      return [sbPrefix, ...decomposeBlock(sb, sbPrefix)];
+    })
+    .reduce(
+      (accFlatSbs, [sbPrefix, sbMainBlock, sbFlatSbs]) => ({
+        ...accFlatSbs,
+        [sbPrefix as string]: sbMainBlock,
+        ...(sbFlatSbs as Record<string, Block>),
+      }),
+      {}
+    );
+  return [mainBlock, flatSubblocks];
+};
+
+/**
+ * Class representing a block directory, providing functionality for adding
+ * blocks and block identifier resolution.
+ */
+class Directory {
+  #blocks: Record<string, Block>;
+
+  constructor() {
+    this.#blocks = {};
+  }
+
+  /**
+   * Adds a block into the directory.
+   *
+   * This function handles the decomposition of blocks into properties and
+   * subblocks. The parent block and each of its subblocks are added to the
+   * block directory, with an appropriate prefix prepended to the block
+   * identifiers.
+   *
+   * For example, by passing in `prefix` as `"foobar"` and `block` as a block
+   * with subblocks `foo` and `bar`, three blocks will be added to the directory
+   * with identifiers `foobar`, `foobar/foo` and `foobar/bar`.
+   *
+   * **Note that each block in the directory will have its subblocks stripped
+   * away. If it is necessary to access a subblock, the [[find]] function should
+   * be used with `prefix` set to the identifier of the parent block and `id`
+   * set to the (partial) identifier of the subblock.**
+   *
+   * @param prefix A prefix string representing the full identifier of the
+   * block.
+   * @param block The block to be added to the directory.
+   */
   addBlock(prefix: string, block: Block): void {
     const [mainBlock, flatSubblocks] = decomposeBlock(block);
-    if (prefix in Object.keys(this.blocks)) {
+    if (prefix in Object.keys(this.#blocks)) {
       throw new Error(`block '${prefix}' already exists`);
     }
 
-    this.blocks[prefix] = mainBlock;
+    this.#blocks[prefix] = mainBlock;
     Object.entries(flatSubblocks).forEach(([name, block]) => {
       const newName = prefix + name;
-      if (newName in Object.keys(this.blocks)) {
+      if (newName in Object.keys(this.#blocks)) {
         throw new Error(`block '${prefix}' already exists`);
       }
-      this.blocks[newName] = block as Block;
+      this.#blocks[newName] = block as Block;
     });
   }
 
+  /**
+   * Finds a block in the directory.
+   *
+   * This function attempts to find a block with the exact same full identifier
+   * as `id`. If no such block exists, it will search for a block with `id`
+   * directly under `prefix`.
+   *
+   * For example, if `prefix` is set to `"abc"` and `id` to `"ghi"`, the
+   * function will check if either of the blocks `ghi` or `abc/ghi` exist.
+   *
+   * @param prefix A prefix string representing the full identifier of the
+   * parent block, if any.
+   * @param id The identifier of the desired block. This may either be a full
+   * identifier or a partial identifier under the parent block.
+   * @return A pair with the first element as the full identifier of the desired
+   * block and the second element as the desired block itself.
+   */
   find(prefix: string, id: BlockId): [BlockId, Block] {
     const prefixedId = prefix === "" ? id : [prefix, id].join("/");
-    if (Object.keys(this.blocks).includes(id)) {
-      return [id, this.blocks[id]];
-    } else if (Object.keys(this.blocks).includes(prefixedId)) {
-      return [prefixedId, this.blocks[prefixedId]];
+    if (Object.keys(this.#blocks).includes(id)) {
+      return [id, this.#blocks[id]];
+    } else if (Object.keys(this.#blocks).includes(prefixedId)) {
+      return [prefixedId, this.#blocks[prefixedId]];
     } else {
       throw new Error(
         id !== prefixedId
